Don't remove mappings when using a custom handler
When changing `g:ag_*handler` it would only make sense to remove the
quickfix/location list mappings if the new handler didn't use one of
those lists.

Otherwise removing the mappings is (imo) a separate decision from
writing your own handler, so it would be best not to mix the two.