Use --column for older versions of ag
Close #82

--vimgrep was introduced in
https://github.com/ggreer/the_silver_searcher/commit/eb163c116b0c3d2e973c4223ab87ed9499083d3b
which is in versions 0.25.0 and up. We use it because if you have one
buffer open and search using :AgBuffer, you get different output with
--column than you would if you didn't search through just that one file.
See https://github.com/rking/ag.vim/pull/80#issuecomment-69460905 for
more discussion on this.

This can probably be removed eventually, I'm not sure how long it takes
package managers to upgrade their versions of ag.