:AgBuffer fix for legacy --column users
Prior to the `--vimgrep` option being added, it was very hard to get Ag
to output file names consistently if only one file was searched rather
than multiple files or a directory [1].

This meant that if you ran an `:AgBuffer` search when you had only
opened one buffer in your Vim session, quickfix navigation would be
broken for your results.

The changed incantation of `ag` in this commit is the best behavior I
was able to get out of earlier Ag versions, after a long saga. It has
one minor drawback: it results in empty newlines between results for
different files, where a heading might otherwise be. This is a small
cosmetic problem--Vim's quickfix parses it just fine and skips over
these on `:cnext` without any glitch or extra keystrokes. This seems a
worthwhile exchange for the outright bug of the single-buffer case.

References d42171e19

[1]: https://github.com/ggreer/the_silver_searcher/issues/361